add modelAPI to pedestian_tracker_demo (#2654)
* add modelAPI to pedestian_tracker_demo last changes

* mist

* mist 2

* add model API to pedestrian_tracker_demo with changes

* solve mistakes

* solve mistakes 3

* solve mistakes 4

* add key person_label and tests

* missing whitespace

* mist

* mist 2

* add model API to pedestrian_tracker_demo with changes

* solve mistakes

* solve mistakes 3

* solve mistakes 4

* add key person_label and tests

* missing whitespace

* cases.py

* update readme.md, main.cpp, pedestrian_tracker_demo.hpp

* cases,py

* Update demos/pedestrian_tracker_demo/cpp/README.md


* back to detection wrapper

* update cases.py

* updates cases.py

* update README.md

* fix warning and stylish

* one more warning fix

* fix yet another warning

+++ demos/pedestrian_tracker_demo/cpp/include/utils.hpp
@@ -1,4 +1,4 @@
-// Copyright (C) 2018-2019 Intel Corporation
+// Copyright (C) 2018-2021 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 //
 
@@ -24,7 +24,7 @@
 ///
 struct DetectionLogEntry {
     TrackedObjects objects;  ///< Detected objects.
-    int frame_idx;           ///< Processed frame index (-1 if N/A).
+    int64_t frame_idx;           ///< Processed frame index (-1 if N/A).
     double time_ms;          ///< Frame processing time in ms (-1 if N/A).
 
     ///
