+++ demos/pedestrian_tracker_demo/cpp/src/tracker.cpp
@@ -1,4 +1,4 @@
-// Copyright (C) 2018-2019 Intel Corporation
+// Copyright (C) 2018-2021 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 //
 
@@ -35,7 +35,7 @@ DetectionLog ConvertTracksToDetectionLog(const ObjectTracks& tracks) {
     DetectionLog log;
 
     // Combine detected objects by respective frame indices.
-    std::map<int, TrackedObjects> objects;
+    std::map<int64_t, TrackedObjects> objects;
     for (const auto& track : tracks)
         for (const auto& object : track.second) {
             auto itr = objects.find(object.frame_idx);
